requirements.txt
numpy==1.14.0
moviepy==0.2.3.2
Pillow==5.0.0
